2_Computer_Science/Artificial_Intelligence_Practice/expr/code/requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.4.0
scipy>=1.7.0
